grants pages revisions and fixes
