+++ apps/fpes/ui/fpes-frontend/src/main.jsx
@@ -2,9 +2,21 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { LocalizationProvider } from "@mui/x-date-pickers";
+import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <App sx={{flexGrow: 1, width:"100%"}}/>
   </StrictMode>,
 )
+
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <React.StrictMode>
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <App />
+    </LocalizationProvider>
+  </React.StrictMode>
+);
